+++ tests/Catalog/OptionTest.php
@@ -3,7 +3,7 @@
  * @copyright Copyright (c) Wizacha
  * @license Proprietary
  */
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Wizaplace\SDK\Tests\Catalog;
 
@@ -35,6 +35,12 @@ final class OptionTest extends TestCase
                     'name' => '21',
                     'image' => [
                         'id' => 1234,
+                        'urls' => [
+                            'original' => '//wathev.er/path/to/an/image.png',
+                            'large' => '//wathev.er/path/to/an/image.png',
+                            'medium' => '//wathev.er/path/to/an/image.png',
+                            'small' => '//wathev.er/path/to/an/image.png',
+                        ],
                     ],
                 ],
             ],
@@ -65,7 +71,13 @@ final class OptionTest extends TestCase
             "id": 4,
             "name": "21",
             "image": {
-                "id": 1234
+                "id": 1234,
+                "urls": {
+                    "original": "\/\/wathev.er\/path\/to\/an\/image.png",
+                    "large": "\/\/wathev.er\/path\/to\/an\/image.png",
+                    "medium": "\/\/wathev.er\/path\/to\/an\/image.png",
+                    "small": "\/\/wathev.er\/path\/to\/an\/image.png"
+                }
             }
         }
     ]